Generic Extension method for OrderQuery
