+++ Repositories/EFCore/Extensions/BookRepositoryExtensions.cs
@@ -37,28 +37,11 @@ namespace Repositories.EFCore.Extensions
             {
                 return books.OrderBy(b => b.Id);
             }
-            var orderParams = orderByQueryString.Trim().Split(',');
-            var propertyInfos = typeof(Book).GetProperties(BindingFlags.Public | BindingFlags.Instance);
+           
+            var orderQuery = OrderQueryBuilder
+                .CreateOrderQuery<Book>(orderByQueryString);
 
-            var orderQueryBuilder = new StringBuilder();
-
-            //title ascending, price descending[,]
-            foreach (var param in orderParams)
-            {
-                if(string.IsNullOrWhiteSpace(param))
-                    continue;
-                var propertyFromQueryName = param.Split(' ')[0];
-                var objectProperty = propertyInfos.FirstOrDefault(pi =>
-                                   pi.Name.Equals(propertyFromQueryName,
-                                   StringComparison.InvariantCultureIgnoreCase));
-                if(objectProperty == null)
-                    continue;
-                var direction = param.EndsWith(" desc") ? "descending" : "ascending";
-                orderQueryBuilder.Append($"{objectProperty.Name.ToString()} {direction},");
-            }
-            var orderQuery = orderQueryBuilder.ToString().TrimEnd(',', ' ');
-
-            if(orderQuery is null)
+            if (orderQuery is null)
             {
                 return books.OrderBy(b => b.Id);
             }
