+++ Repositories/EFCore/Extensions/OrderQueryBuilder.cs
@@ -1,0 +1,41 @@
+﻿using Entities.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Repositories.EFCore.Extensions
+{
+    public static class OrderQueryBuilder
+    {
+        public static String CreateOrderQuery<T>(String orderByQueryString)
+        {
+            var orderParams = orderByQueryString.Trim().Split(',');
+
+            var propertyInfos = typeof(T).GetProperties(BindingFlags.Public | BindingFlags.Instance);
+
+            var orderQueryBuilder = new StringBuilder();
+
+            foreach (var param in orderParams)
+            {
+                if (string.IsNullOrWhiteSpace(param))
+                    continue;
+                var propertyFromQueryName = param.Split(' ')[0];
+                var objectProperty = propertyInfos.FirstOrDefault(pi =>
+                                   pi.Name.Equals(propertyFromQueryName,
+                                   StringComparison.InvariantCultureIgnoreCase));
+                if (objectProperty == null)
+                    continue;
+                //title ascending, price descending[,]
+                var direction = param.EndsWith(" desc") ? "descending" : "ascending";
+                orderQueryBuilder.Append($"{objectProperty.Name.ToString()} {direction},");
+
+              
+            }
+            var orderQuery = orderQueryBuilder.ToString().TrimEnd(',', ' ');
+            return orderQuery;
+        }
+    }
+}
